grammar: distinguish between property and variable assignments
This commit creates a distinction between property and variable
assignments within a declaration.  When the left side of a declaration
is a variable, the token `(variable)` will now appear in the grammar
rather than `(property_name)`.

This helps enforce the separation between Sass and CSS, and allows
the `!global` flag to only appear in variable assignments, while
allowing the `!important` (and Sass-related `!default`) keys to
appear only on property assignments.

This fixes [Serenadeai Issue 11](https://github.com/serenadeai/tree-sitter-scss/issues/11).

\# What

\# Why

\# How

\# Designs

\# Test Steps

\# Other Notes