escape sequences are valid values
\# What

This commit adds `$.escape_sequence` to the list of value `$._value` choices.  This allows
for declarations where the right side is just a single escape, which is common

This might be a bug with tree-sitter-css, but that's not my responsibility right now.

Added an `eslint.config.mjs` configuration file, using ESLint 9.  This is mostly to pacify
IDEs that expect eslint to be configured and available as part of the Javascript development
process.  Turns out there's already an eslint-treesitter-config stock configuration, so I
just used it.

\# How to Test

```
$ tree-sitter generate && tree-sitter test -i 'Variables With Escape Sequence'
$ tree-sitter generate && tree-sitter test -i 'Custom Property With Escape Sequence'
```